FIX: 견적서 리스트 지점, 딜러 명 정보 post 수정

+++ dealers/models.py
@@ -15,8 +15,8 @@ class Dealer(TimeStampModel):
 class Branch(TimeStampModel):
     name       = models.CharField(max_length = 100)
     address    = models.CharField(max_length = 200)
-    latitude   = models.DecimalField(max_digits = 6, decimal_places = 6, default = 0)
-    longitude = models.DecimalField(max_digits = 6, decimal_places = 6, default = 0)
+    latitude   = models.FloatField(default = 0)
+    longitude  = models.FloatField(default = 0)
 
     class Meta:
         db_table = 'branches'
